Moved enums to helpers file
