+++ src/components/helpers/helpers.ts
@@ -1,3 +1,16 @@
+export enum Position {
+  TOP = 'top',
+  RIGHT = 'right',
+  BOTTOM = 'bottom',
+  LEFT = 'left',
+}
+
+export enum Variant {
+  DARK = 'dark',
+  LIGHT = 'light',
+  ERROR = 'error',
+}
+
 export interface MouseEvent extends Event {
   path: Node[];
 }
