--- src/components/zen-tooltip/types.tsx
@@ -1,12 +1,0 @@
-export enum Position {
-  TOP = 'top',
-  RIGHT = 'right',
-  BOTTOM = 'bottom',
-  LEFT = 'left',
-}
-
-export enum Variant {
-  DARK = 'dark',
-  LIGHT = 'light',
-  ERROR = 'error',
-}
